fixed typing issues

+++ chull.hh
@@ -13,7 +13,7 @@ class QuickHull{
     #define iPair std::pair<int, int>
     std::set<iPair> hull;
 public:
-    int counter = 0;
+    long int counter = 0;
     // Returns the side of point p with respect to line
     // joining points p1 and p2.
     int findSide(iPair p1, iPair p2, iPair p)
@@ -41,13 +41,13 @@ public:
 
     // End points of line L are p1 and p2.  side can have value
     // 1 or -1 specifying each of the parts made by the line L
-    void quickHull(iPair a[], int n, iPair p1, iPair p2, int side)
+    void quickHull(iPair a[], long int n, iPair p1, iPair p2, int side)
     {
         int ind = -1;
         int max_dist = 0;
         // finding the point with maximum distance
         // from L and also on the specified side of L.
-        for (int i=0; i<n; i++)
+        for (long int i=0; i<n; i++)
         {
             int temp = lineDist(p1, p2, a[i]);
             if (findSide(p1, p2, a[i]) == side && temp > max_dist)
@@ -72,7 +72,7 @@ public:
         quickHull(a, n, a[ind], p2, -findSide(a[ind], p2, p1));
     }
 
-    void printHull(iPair a[], int n)
+    void printHull(iPair a[], long int n)
     {
         // a[i].second -> y-coordinate of the ith point
         if (n < 3)
@@ -83,8 +83,8 @@ public:
 
         // Finding the point with minimum and
         // maximum x-coordinate
-        int min_x = 0, max_x = 0;
-        for (int i=1; i<n; i++)
+        long int min_x = 0, max_x = 0;
+        for (long int i=1; i<n; i++)
         {
             if (a[i].first < a[min_x].first)
             min_x = i;
@@ -117,4 +117,3 @@ public:
     }
 };
 #endif
-
